requirements.txt
ccxt>=4.4.0,<4.5.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.0
rich>=13.0.0
